requirements.txt
requests>=2.28.0
python-frontmatter>=1.0.0
